- 커스텀 예외 통일 - 필드들 private 접근자로 지정 - controller dto 예외 처리 - 중복된 dto 예외 처리 구문 제거

+++ src/main/java/com/flabedu/small/small/model/Item.java
@@ -1,7 +1,8 @@
 package com.flabedu.small.small.model;
 
-import lombok.*;
+import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class Item {
     private final String name;
     private final String nameEng;
     private final String gender;
-    private final long price;
+    private final BigDecimal price;
     private final String registerUserID;
     private final LocalDateTime registerDate;
     private final String modifiedUserID;
