+++ src/test/java/com/flabedu/small/small/mapper/OrdersItemMapperTest.java
@@ -7,6 +7,7 @@ import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -28,8 +29,8 @@ class OrdersItemMapperTest {
     public void saveOrderDetailTest(){
         List<OrdersItem> inputList = new ArrayList<>();
         var ordersId = 1;
-        var ordersItem1 =new OrdersItem(ordersId, 2, 2, 1, 5000);
-        var ordersItem2 =new OrdersItem(ordersId, 2, 3, 2, 6000);
+        var ordersItem1 =new OrdersItem(ordersId, 2, 2, 1, BigDecimal.valueOf(5000));
+        var ordersItem2 =new OrdersItem(ordersId, 2, 3, 2, BigDecimal.valueOf(6000));
         inputList.add(ordersItem1);
         inputList.add(ordersItem2);
